refactoring

+++ TYVObjCHomework/Source/TYVCarwash/TYVRooms/TYVCarwashRoom/TYVCarwashRoom.h
@@ -10,12 +10,12 @@
 #import "TYVCar.h"
 
 @interface TYVCarwashRoom : TYVRoom
-@property (nonatomic, readonly)                           NSArray  *cars;
-@property (nonatomic, readonly)                           uint8_t  carCount;
-@property (nonatomic, readonly, getter=isFullForHuman)    BOOL     fullForHuman;
-@property (nonatomic, readonly, getter=isFullForCar)      BOOL     fullForCar;
+@property (nonatomic, readonly)                           NSArray       *cars;
+@property (nonatomic, readonly)                           NSUInteger    carCount;
+@property (nonatomic, readonly, getter=isFullForHuman)    BOOL          fullForHuman;
+@property (nonatomic, readonly, getter=isFullForCar)      BOOL          fullForCar;
 
-- (instancetype)initWithHumanCount:(uint8_t)humanCount carCount:(uint8_t)carCount;
+- (instancetype)initWithHumanCount:(NSUInteger)humanCount carCount:(NSUInteger)carCount;
 
 - (void)addCar:(TYVCar *)aCar;
